datatype: add opt_desc/desc equivalence test and shared corpus
Introduce a reusable, public-MPI-only datatype corpus
(ompi/test/datatype/datatype_corpus.[ch]) that builds and commits a
curated set of committed datatypes, each paired with an independent
by-hand pack/unpack reference and a set of shape traits.  The corpus is
consumer-agnostic: it describes each datatype but knows nothing about
how a given test selects, drives, or reports on them, so more than one
test can share a single definition of each construction.

Add opt_desc_equiv, which drives every corpus entry through both the
optimized descriptor and a generic walk over the canonical descriptor,
across a sweep of fragment sizes, and asserts the two packed streams are
byte-identical and both round-trip correctly.  This catches optimizer
rewrites that would change the on-the-wire layout.

Refactor to_self to consume the corpus instead of carrying its own
builders and by-hand baselines; it keeps its historical --data selection
by matching each corpus entry's name against its own option table.  The
builders (including the DDTBench-derived shapes) move verbatim into the
corpus, and a few adversarial shapes -- single-iteration, zero/negative
extent, overlap, and mixed-type promotion -- are added to stress the
optimizer and positioning edges.

Also fix OPAL_DATATYPE_SAFEGUARD_POINTER to bound negative-extent
datatypes correctly.  Instances may stride backward, so the region
touched by COUNT copies is the union over all of them: fold a negative
span into the lower bound and a positive span into the upper bound
rather than assuming the last copy sits at the highest address.  Without
this the corpus's adversarial negative-extent shape trips false-positive
debug warnings.
