+++ ompi/test/datatype/Makefile.am
@@ -41,7 +41,8 @@ TESTS = \
         partial \
         ompi_datatype_bigcount \
         mpi_datatype_bigcount \
-        sizedtypes
+        sizedtypes \
+        opt_desc_equiv
 
 MPI_CHECKS = to_self pack_description_sweep reduce_local
 
@@ -84,9 +85,13 @@ position_noncontig_SOURCES = position_noncontig.c
 position_noncontig_LDFLAGS = $(common_ldflags)
 position_noncontig_LDADD = $(common_ldadd)
 
-to_self_SOURCES = to_self.c
+to_self_SOURCES = to_self.c datatype_corpus.c datatype_corpus.h
 to_self_LDFLAGS = $(common_ldflags)
 to_self_LDADD = $(common_ldadd)
+
+opt_desc_equiv_SOURCES = opt_desc_equiv.c datatype_corpus.c datatype_corpus.h
+opt_desc_equiv_LDFLAGS = $(common_ldflags)
+opt_desc_equiv_LDADD = $(common_ldadd)
 
 pack_description_sweep_SOURCES = pack_description_sweep.c
 pack_description_sweep_LDFLAGS = $(common_ldflags)
